SimpleLottery: Moving algo.rs in proof.rs implementation block
